+++ firebase-core/src/main/kotlin/com/androidhuman/rxfirebase2/core/SimpleDisposable.kt
@@ -15,5 +15,5 @@ abstract class SimpleDisposable : Disposable {
         }
     }
 
-    abstract protected fun onDispose()
+    protected abstract fun onDispose()
 }